Refactor paragraph functions into modules
