+++ importers/src/bible/osis/tests/functions/helpers.functions.test.ts
@@ -1,6 +1,7 @@
-import { getEmptyParagraph, getEmptySection } from './../utils';
+import { getEmptySection } from './../utils';
 import { ParserContext } from './../../entities/ParserContext';
-import { getCurrentContainer, isBeginningOfParagraph } from '../../functions/helpers.functions';
+import { getCurrentContainer } from '../../functions/helpers.functions';
+import { getEmptyParagraph } from '../../functions/paragraphs.functions';
 
 describe('getCurrentContainer', () => {
     it('throws error if no valid current container', () => {
@@ -11,13 +12,5 @@ describe('getCurrentContainer', () => {
         const context = new ParserContext()
         context.contentContainerStack.push(getEmptyParagraph(), getEmptySection())
         expect(getCurrentContainer(context).type).toBe('section')
-    })
-})
-
-describe('isBeginningOfParagraph', () => {
-    it('identifies empty paragraphs', () => {
-        const context = new ParserContext()
-        context.contentContainerStack.push(getEmptyParagraph())
-        expect(isBeginningOfParagraph(context)).toBe(true)
     })
 })
